requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.12.7
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
crispy-bootstrap5==0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
dash==2.8.1
dash-bootstrap-components==1.4.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
directory-tree-generator==0.0.1
distlib==0.3.5
Django==3.2.18
django-crispy-forms==1.14.0
django-filter==23.2
django-plotly-dash==2.1.3
djangorestframework==3.14.0
dpd-components==0.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.8.0
Flask==2.2.3
flatbuffers==22.12.6
fontawesomefree==6.1.2
fonttools==4.38.0
gast==0.5.3
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
imutils==0.5.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mediapipe==0.8.3.1
msgpack==1.0.4
multitasking==0.0.11
mysql-connector-python==8.0.29
mysqlclient==2.1.1
networkx==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.13.1
pluggy==1.0.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrealsense2==2.53.1.4623
pytest==7.2.0
python-dateutil==2.8.2
python-osc==1.8.1
pytz==2022.6
PyWavelets==1.3.0
pywin32==305
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sqlparse==0.4.2
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.5.4
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.13
virtualenv==20.16.3
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.14.1
zipfile36==0.1.3
zipp==3.11.0
